adding middleware usage
Previous functionality had lack of flexibility with middleware. This meant significant limitations with how you can allow guest and auth users to view docs. If auth was set to true (now deprecated), guests could not see your docs. But often times it's nice to have a subset of docs that are public, a subset that are for users, and at least one more subset for, perhaps, admins.